# sandbox/api/openapi.yaml
openapi: 3.0.3
info:
  description: |
    API for communication suppliers to integrate with NHS Notify.

    This specification represents the in-development 'next' version of the API schema
    and should be treated as unstable.
  title: NHS Notify Supplier API
  version: next
servers:
- description: Local development server
  url: http://127.0.0.1:3000
- description: Public sandbox
  url: https://sandbox-server.nhs.uk/nhs-notify-supplier-api
...
paths:
  /letters:
    get:
      description: The key use of this endpoint is to query letters which are ready
        to be printed
      operationId: listLetters
      parameters:
      - description: "Unique request identifier, in the format of a GUID"
        explode: false
        in: header
        name: X-Request-ID
        required: true
        schema:
          type: string
        style: simple
      - description: |-
          An optional ID which you can use to track transactions across multiple systems. It can take any value, but we recommend avoiding `.` characters. If not provided in the request, NHS Notify will default to a system generated ID in its place.
          The ID will be returned in a response header.
        explode: false
        in: header
        name: X-Correlation-ID
        required: false
        schema:
          type: string
        style: simple
      - description: |-
          Limit maximum number of items to be fetched
        explode: true
        in: query
        name: limit
        required: false
        schema:
          example: 10
          type: number
        style: form
      responses:
        "200":
          content:
            application/vnd.api+json:
              examples:
                get-letters:
                  value:
                    data:
                    - attributes:
                        specificationId: 2WL5eYSWGzCHlGmzNxuqVusPxDg
                        status: PENDING
                      id: 2WL5eYSWGzCHlGmzNxuqVusPxDg
                      type: Letter
              schema:
                $ref: "#/components/schemas/listLetters_200_response"
          description: List of letters to process
        "404":
          content:
            application/vnd.api+json:
              examples:
                error-not-found:
                  value:
                    errors:
                    - code: NOTIFY_RESOURCE_NOT_FOUND
                      detail: No resource found with that ID
                      id: rrt-1931948104716186917-c-geu2-10664-3111479-3.0
                      links:
                        about: https://digital.nhs.uk/developer/api-catalogue/nhs-notify-supplier
                      status: "404"
                      title: Resource not found
              schema:
                $ref: "#/components/schemas/listLetters_404_response"
          description: Resource not found
        "429":
          content:
            application/vnd.api+json:
              examples:
                error-too-many-requests:
                  value:
                    errors:
                    - code: NOTIFY_QUOTA
                      detail: You have made too many requests. Please try again later.
                      id: rrt-1931948104716186917-c-geu2-10664-3111479-3.0
                      links:
                        about: https://digital.nhs.uk/developer/api-catalogue/nhs-notify-supplier
                      status: "429"
                      title: Too many requests
              schema:
                $ref: "#/components/schemas/listLetters_404_response"
          description: Too many requests
        "500":
          content:
            application/vnd.api+json:
              examples:
                error-server-error:
                  value:
                    errors:
                    - code: NOTIFY_INTERNAL_SERVER_ERROR
                      detail: "An unexpected error occurred, please contact support"
                      id: rrt-1931948104716186917-c-geu2-10664-3111479-3.0
                      links:
                        about: https://digital.nhs.uk/developer/api-catalogue/nhs-notify-supplier
                      status: "500"
                      title: Internal Server Error
              schema:
                $ref: "#/components/schemas/listLetters_404_response"
          description: Server error
      summary: Get a list of letters
      tags:
      - letter
      x-eov-operation-handler: controllers/LetterController
    post:
      operationId: postLetters
      parameters:
      - description: "Unique request identifier, in the format of a GUID"
        explode: false
        in: header
        name: X-Request-ID
        required: true
        schema:
          type: string
        style: simple
      - description: |-
          An optional ID which you can use to track transactions across multiple systems. It can take any value, but we recommend avoiding `.` characters. If not provided in the request, NHS Notify will default to a system generated ID in its place.
          The ID will be returned in a response header.
        explode: false
        in: header
        name: X-Correlation-ID
        required: false
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/vnd.api+json:
            examples:
              post-letter-request:
                value:
                  data:
                  - attributes:
                      reasonCode: 100
                      reasonText: failed validation
                      requestedProductionStatus: ACTIVE
                      specificationId: 2WL5eYSWGzCHlGmzNxuqVusPxDg
                      status: REJECTED
                    id: 2WL5eYSWGzCHlGmzNxuqVusPxDg
                    type: Letter
            schema:
              $ref: "#/components/schemas/postLetters_request"
        required: true
      responses:
        "200":
          content:
            application/vnd.api+json:
              examples:
                post-letters-responses:
                  value:
                    data:
                    - attributes:
                        reasonCode: 100
                        reasonText: failed validation
                        requestedProductionStatus: ACTIVE
                        specificationId: 2WL5eYSWGzCHlGmzNxuqVusPxDg
                        status: REJECTED
                      id: 2WL5eYSWGzCHlGmzNxuqVusPxDg
                      type: Letter
              schema:
                $ref: "#/components/schemas/postLetters_200_response"
          description: Letter Resources Updated successfully
        "404":
          content:
            application/vnd.api+json:
              examples:
                error-not-found:
                  value:
                    errors:
                    - code: NOTIFY_RESOURCE_NOT_FOUND
                      detail: No resource found with that ID
                      id: rrt-1931948104716186917-c-geu2-10664-3111479-3.0
                      links:
                        about: https://digital.nhs.uk/developer/api-catalogue/nhs-notify-supplier
                      status: "404"
                      title: Resource not found
              schema:
                $ref: "#/components/schemas/listLetters_404_response"
          description: Resource not found
        "429":
          content:
            application/vnd.api+json:
              examples:
                error-too-many-requests:
                  value:
                    errors:
                    - code: NOTIFY_QUOTA
                      detail: You have made too many requests. Please try again later.
                      id: rrt-1931948104716186917-c-geu2-10664-3111479-3.0
                      links:
                        about: https://digital.nhs.uk/developer/api-catalogue/nhs-notify-supplier
                      status: "429"
                      title: Too many requests
              schema:
                $ref: "#/components/schemas/listLetters_404_response"
          description: Too many requests
        "500":
          content:
            application/vnd.api+json:
              examples:
                error-server-error:
                  value:
                    errors:
                    - code: NOTIFY_INTERNAL_SERVER_ERROR
                      detail: "An unexpected error occurred, please contact support"
                      id: rrt-1931948104716186917-c-geu2-10664-3111479-3.0
                      links:
                        about: https://digital.nhs.uk/developer/api-catalogue/nhs-notify-supplier
                      status: "500"
                      title: Internal Server Error
              schema:
                $ref: "#/components/schemas/listLetters_404_response"
          description: Server error
      summary: Update the status of multiple letters
      tags:
      - letter
      x-eov-operation-handler: controllers/LetterController
...
components:
  schemas:
    listLetters_200_response_links:
      additionalProperties: false
      description: Contains links to other data pages
      properties:
        last:
          description: URI of final page of data
          example: https://dev.api.service.nhs.uk/nhs-notify-supplier/letters?status=PENDING&page=10
          format: uri
          type: string
        next:
          description: URI of next page of data
          example: https://dev.api.service.nhs.uk/nhs-notify-supplier/letters?status=PENDING&page=5
          format: uri
          type: string
        prev:
          description: URI of prev page of data
          example: https://dev.api.service.nhs.uk/nhs-notify-supplier/letters?status=PENDING&page=7
          format: uri
          type: string
        self:
          description: URI of current page of data
          example: https://dev.api.service.nhs.uk/nhs-notify-supplier/letters?status=PENDING&page=6
          format: uri
          type: string
      type: object
    postLetters_200_response:
      properties:
        data:
          items:
            $ref: "#/components/schemas/listLetters_200_response_data_inner"
          type: array
      type: object
...
    listLetters_200_response_data_inner_attributes:
      properties:
        specificationId:
          type: string
        status:
          default: PENDING
          description: The supplier status of an individual letter
          enum:
          - PENDING
          - ACCEPTED
          - REJECTED
          - PRINTED
          - ENCLOSED
          - CANCELLED
          - DISPATCHED
          - DELIVERED
          - FAILED
          - RETURNED
          - DESTROYED
          - FORWARDED
          type: string
        requestedProductionStatus:
          description: A requested status for the production of a letter
          enum:
          - ACTIVE
          - HOLD
          - CANCEL
          title: ProductionStatus
          type: string
        reasonCode:
          description: Reason code for the given status
          type: number
        reasonText:
          description: Reason text for the given status
          type: string
      required:
      - specificationId
      - status
      type: object
    listLetters_404_response:
      properties:
        data:
          $ref: "#/components/schemas/listLetters_404_response_data"
      type: object
...
    listLetters_404_response_data:
      properties:
        errors:
          items:
            $ref: "#/components/schemas/listLetters_404_response_data_errors_inner"
          type: array
      type: object
    listLetters_404_response_data_errors_inner_links:
      properties:
        about:
          format: uri
          type: string
      type: object
...
    listLetters_200_response:
      properties:
        data:
          items:
            $ref: "#/components/schemas/listLetters_200_response_data_inner"
          type: array
        links:
          $ref: "#/components/schemas/listLetters_200_response_links"
      type: object
...
    postLetters_request_data_inner_attributes:
      properties:
        specificationId:
          type: string
        status:
          default: PENDING
          description: The supplier status of an individual letter
          enum:
          - PENDING
          - ACCEPTED
          - REJECTED
          - PRINTED
          - ENCLOSED
          - CANCELLED
          - DISPATCHED
          - DELIVERED
          - FAILED
          - RETURNED
          - DESTROYED
          - FORWARDED
          type: string
        requestedProductionStatus:
          description: |-
            The requested production status for this letter.
            May only be set by NHS Notify.
          enum:
          - ACTIVE
          - HOLD
          - CANCEL
          title: ProductionStatus
          type: string
        reasonCode:
          description: Reason code for the given status
          type: number
        reasonText:
          description: Reason text for the given status
          type: string
      type: object
    listLetters_200_response_data_inner:
      properties:
        type:
          type: string
        id:
          type: string
        attributes:
          $ref: "#/components/schemas/listLetters_200_response_data_inner_attributes"
      type: object
    postLetters_request_data_inner:
      properties:
        type:
          type: string
        id:
          type: string
        attributes:
          $ref: "#/components/schemas/postLetters_request_data_inner_attributes"
      type: object
...
    postLetters_request:
      properties:
        data:
          items:
            $ref: "#/components/schemas/postLetters_request_data_inner"
          type: array
      type: object
...
    listLetters_404_response_data_errors_inner:
      properties:
        id:
          type: string
        code:
          enum:
          - NOTIFY_INTERNAL_SERVER_ERROR
          - NOTIFY_INVALID_REQUEST
          - NOTIFY_LETTER_NOT_FOUND
          type: string
        links:
          $ref: "#/components/schemas/listLetters_404_response_data_errors_inner_links"
        status:
          type: string
        title:
          type: string
        detail:
          type: string
      type: object
...
  securitySchemes:
    nhs-login-p0:
      description: |
        # NHS Login P0 authentication

        This is a user-restricted authentication pattern for Patient Access to
        NHS Digital APIs.  For a full description of our authentication and
        authorization patterns please see the [Security and authorisation page](https://digital.nhs.uk/developer/guides-and-documentation/security-and-authorisation).
      flows:
        authorizationCode:
          authorizationUrl: /oauth2/authorize
          scopes: {}
          tokenUrl: /oauth2/token
        x-tokenExchange:
          tokenUrl: /oauth2/token
          scopes: {}
      type: oauth2
    app-level0:
      description: |
        # Application level 0 authentication

        This is an application-restricted authentication pattern for access
        to NHS England APIs.  For a full description of our authentication
        and authorization patterns please see the [Security and authorisation page](https://digital.nhs.uk/developer/guides-and-documentation/security-and-authorisation).
      in: header
      name: apikey
      type: apiKey
...